LayoutTests/editing/editability/attr-missing-ancestor-false.html

<html>
<head>
<link rel="stylesheet" href="../../fast/js/resources/js-test-style.css">
<script src="../../fast/js/resources/js-test-pre.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></p>
<div id="result"></div>
<div id="div" contentEditable="false">
  <div>
    <div>
      <p id="p"></p>
    </div>
  </div>
</div>
<script>
description('When contentEditable is missing, it maps to "inherit" state and the element is NOT editable if its nearest ancestor is NOT.')

shouldBe('document.getElementById("p").hasAttribute("contentEditable")', 'false');
shouldBe('document.getElementById("p").contentEditable', '"inherit"');
debug("FIXME: contentEditable attribute should be inherit if missing. https://bugs.webkit.org/show_bug.cgi?id=52056");
shouldBe('document.getElementById("p").isContentEditable', 'false');
shouldBe('window.getComputedStyle(p, "").getPropertyValue("-webkit-user-modify")', '"read-only"');

document.getElementById("div").style.display= 'none';

window.successfullyParsed = true;
</script>
<script src="../../fast/js/resources/js-test-post.js"></script>
</body>
</html>
